committed

+++ src/main/java/com/example/demo/repositry/RepoStudent.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entities.Student;
 
-public interface RepoStudent extends JpaRepository<Student, Integer> {
+public interface RepoStudent extends JpaRepository<Student, Long> {
 
 }
